Backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings>=2
opencv-python
sqlalchemy
pymysql
paramiko
passlib[bcrypt]
python-jose[cryptography]
slowapi
python-dotenv
